requirements.txt
Django==1.11.4
DjangoRestless==0.0.10
pytz==2017.2
six==1.10.0
